+++ insteon_mqtt/device/Remote.py
@@ -5,6 +5,7 @@
 #===========================================================================
 from ..CommandSeq import CommandSeq
 from .. import log
+from .. import on_off
 from ..Signal import Signal
 from .Base import Base
 
@@ -51,10 +52,6 @@ class Remote(Base):
                  see if the database is current.  Reloads the modem database
                  if needed.  This will emit the current state as a signal.
     """
-
-    on_codes = [0x11, 0x12, 0x21, 0x23]  # on, fast on, instant on, manual on
-    off_codes = [0x13, 0x14, 0x22]  # off, fast off, instant off
-
     def __init__(self, protocol, modem, address, name, num_button):
         """Constructor
 
@@ -71,7 +68,8 @@ class Remote(Base):
         self.num = num_button
         self.type_name = "mini_remote_%d" % self.num
 
-        self.signal_pressed = Signal()  # (Device, int group, bool on)
+        # (Device, int group, bool on, on_off.Mode mode)
+        self.signal_pressed = Signal()
 
     #-----------------------------------------------------------------------
     def pair(self, on_done=None):
@@ -133,7 +131,7 @@ class Remote(Base):
         Args:
           msg:   (InptStandard) Broadcast message from the device.
         """
-        is_on = None
+        is_on, mode = None, on_off.Mode.NORMAL
         cmd = msg.cmd1
 
         # ACK of the broadcast - ignore this.
@@ -141,15 +139,11 @@ class Remote(Base):
             LOG.info("Remote %s broadcast ACK grp: %s", self.addr, msg.group)
             return
 
-        # On command.  0x11: on, 0x12: on fast
-        elif cmd in Remote.on_codes:
-            LOG.info("Remote %s broadcast ON grp: %s", self.addr, msg.group)
-            is_on = True
-
-        # Off command. 0x13: off, 0x14: off fast
-        elif cmd in Remote.off_codes:
-            LOG.info("Remote %s broadcast OFF grp: %s", self.addr, msg.group)
-            is_on = False
+        # On/off command codes.
+        elif on_off.Mode.is_valid(msg.cmd1):
+            is_on, mode = on_off.Mode.decode(msg.cmd1)
+            LOG.info("Remote %s broadcast grp: %s on: %s mode: %s", self.addr,
+                     msg.group, is_on, mode)
 
         # Starting manual increment (cmd2 0x00=up, 0x01=down)
         elif cmd == 0x17:
@@ -165,7 +159,7 @@ class Remote(Base):
 
         # Notify others that the button was pressed.
         if is_on is not None:
-            self.signal_pressed.emit(self, msg.group, is_on)
+            self.signal_pressed.emit(self, msg.group, is_on, mode)
 
         # This will find all the devices we're the controller of for
         # this group and call their handle_group_cmd() methods to