+++ insteon_mqtt/mqtt/Switch.py
@@ -5,6 +5,7 @@
 #===========================================================================
 from .. import log
 from .MsgTemplate import MsgTemplate
+from .. import on_off
 
 LOG = log.get_logger()
 
@@ -19,6 +20,33 @@ class Switch:
     Some classes that can act like a switch can inherit from this
     class to use the same MQTT templates (see Dimmer).
     """
+    #-----------------------------------------------------------------------
+    @classmethod
+    def parse_json(cls, data):
+        """TODO: doc
+        """
+        cmd = data.get('cmd')
+        if cmd == 'on':
+            is_on = True
+        elif cmd == 'off':
+            is_on = False
+        else:
+            raise Exception("Invalid on/off command input '%s'" % cmd)
+
+        # If mode is present, use that to specify normal/fast/instant.
+        # Otherwise look for individual keywords.
+        if 'mode' in data:
+            mode = on_off.Mode(data.get('mode', 'normal'))
+        else:
+            mode = on_off.Mode.NORMAL
+            if data.get('fast', False):
+                mode = on_off.Mode.FAST
+            elif data.get('instant', False):
+                mode = on_off.Mode.INSTANT
+
+        return is_on, mode
+
+    #-----------------------------------------------------------------------
     def __init__(self, mqtt, device, handle_active=True):
         """Constructor
 
@@ -36,20 +64,17 @@ class Switch:
         # Output state change reporting template.
         self.msg_state = MsgTemplate(
             topic='insteon/{{address}}/state',
-            payload='{{on_str.lower()}}',
-            )
+            payload='{{on_str.lower()}}')
 
         # Input on/off command template.
         self.msg_on_off = MsgTemplate(
             topic='insteon/{{address}}/set',
-            payload='{ "cmd" : "{{value.lower()}}" }',
-            )
+            payload='{ "cmd" : "{{value.lower()}}" }')
 
         # Input scene on/off command template.
         self.msg_scene_on_off = MsgTemplate(
             topic='insteon/{{address}}/scene',
-            payload='{ "cmd" : "{{value.lower()}}" }',
-            )
+            payload='{ "cmd" : "{{value.lower()}}" }')
 
         # Receive notifications from the Insteon device when it changes.
         if handle_active:
@@ -89,7 +114,7 @@ class Switch:
           qos:    The quality of service to use.
         """
         topic = self.msg_on_off.render_topic(self.template_data())
-        link.subscribe(topic, qos, self.handle_set)
+        link.subscribe(topic, qos, self.handle_on_off)
 
         topic = self.msg_scene_on_off.render_topic(self.template_data())
         link.subscribe(topic, qos, self.handle_scene)
@@ -108,7 +133,7 @@ class Switch:
         link.unsubscribe(topic)
 
     #-----------------------------------------------------------------------
-    def template_data(self, is_active=None):
+    def template_data(self, is_on=None, mode=on_off.Mode.NORMAL):
         """TODO: doc
         """
         # Set up the variables that can be used in the templates.
@@ -118,14 +143,17 @@ class Switch:
                      else self.device.addr.hex,
             }
 
-        if is_active is not None:
-            data["on"] = 1 if is_active else 0
-            data["on_str"] = "on" if is_active else "off"
+        if is_on is not None:
+            data["on"] = 1 if is_on else 0
+            data["on_str"] = "on" if is_on else "off"
+            data["mode"] = str(mode)
+            data["fast"] = 1 if mode == on_off.Mode.FAST else 0
+            data["instant"] = 1 if mode == on_off.Mode.INSTANT else 0
 
         return data
 
     #-----------------------------------------------------------------------
-    def handle_active(self, device, is_active):
+    def handle_active(self, device, is_on, mode=on_off.Mode.NORMAL):
         """Device active on/off callback.
 
         This is triggered via signal when the Insteon device goes
@@ -136,14 +164,15 @@ class Switch:
           device:   (device.Base) The Insteon device that changed.
           is_active (bool) True for on, False for off.
         """
-        LOG.info("MQTT received active change %s = %s", device.label,
-                 is_active)
+        LOG.info("MQTT received active change %s = %s %s", device.label,
+                 is_on, mode)
 
-        data = self.template_data(is_active)
+        data = self.template_data(is_on, mode)
+
         self.msg_state.publish(self.mqtt, data)
 
     #-----------------------------------------------------------------------
-    def handle_set(self, client, data, message):
+    def handle_on_off(self, client, data, message):
         """TODO: doc
         """
         LOG.debug("Switch message %s %s", message.topic, message.payload)
@@ -153,21 +182,12 @@ class Switch:
         LOG.info("Switch input command: %s", data)
 
         try:
-            cmd = data.get('cmd')
-            if cmd == 'on':
-                is_on = True
-            elif cmd == 'off':
-                is_on = False
-            else:
-                raise Exception("Invalid switch cmd input '%s'" % cmd)
-
-            instant = bool(data.get('instant', False))
+            # Tell the device to update it's state.
+            is_on, mode = Switch.parse_json(data)
+            self.device.set(level=is_on, mode=mode)
         except:
             LOG.exception("Invalid switch command: %s", data)
             return
-
-        # Tell the device to update it's state.
-        self.device.set(is_on, instant=instant)
 
     #-----------------------------------------------------------------------
     def handle_scene(self, client, data, message):
@@ -180,20 +200,14 @@ class Switch:
         LOG.info("Switch input command: %s", data)
 
         try:
-            cmd = data.get('cmd')
-            if cmd == 'on':
-                is_on = True
-            elif cmd == 'off':
-                is_on = False
-            else:
-                raise Exception("Invalid switch cmd input '%s'" % cmd)
-
+            is_on, _mode = Switch.parse_json(data)
             group = int(data.get('group', 0x01))
+
+            # Tell the device to trigger the scene command.
+            self.device.scene(is_on, group)
+
         except:
             LOG.exception("Invalid switch command: %s", data)
             return
-
-        # Tell the device to trigger the scene command.
-        self.device.scene(is_on, group)
 
     #-----------------------------------------------------------------------